refactor(web): move project settings page into features/projects
Preserve history with git mv. Replace the pages-only null gate with
NoDataOrLoading so the file passes @repo/no-null-render under features/.
Route shim lands in a follow-up commit.
